% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{Box-Cox Transformation}
\usage{
boxcox(x, lambda = 0, plot_it = FALSE)
}
\arguments{
\item{x}{a vector to which you are applying the box-cox transformation. This value must be positive.}

\item{lambda}{the power parameter you are applying to a vector. A default of 0 is used.}
}
\value{
\itemize{
   \item if lambda = 0, ln(x)
   \item for all other lambda, (x^lambda-1)/lambda
}
}
\description{
Apply box-cox transformation to move towards normality
}
