% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frac_pow.R
\name{frac.pow}
\alias{frac.pow}
\title{Generic Fractional Power Function}
\usage{
frac.pow(x, r = 2, plot_it = FALSE)
}
\arguments{
\item{x}{Apply a fractional power to this vector}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.}

\item{a}{Raise \code{x} by this fractional power}
}
\value{
The vector \code{x}, raised to the fractional power of \code{a}.
}
\description{
I want users to have the freedom to apply any fractional power they desire, so I am exporting it.
}
