% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.frac.pow.R
\name{squarer}
\alias{squarer}
\title{Apply a useful fractional power}
\usage{
squarer(x, plot_it = FALSE)
}
\arguments{
\item{x}{Get the square or cubic root of this value}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item square root (for \code{squarer})
     \item cube root (for \code{cubicr})
}
 of \code{x}.
}
\description{
As the title states, these functions take the cubic and square root of vectors
}
\details{
To find out more, please look into the \code{\link{frac.pow}}
function that these functions depend on.
}
\examples{
squarer(tenvec)
cubicr(64)
}
